requirements.txt
torch>=2.0.0
torchvision
transformers>=4.30.0
diffusers>=0.21.0
numpy
Pillow
scipy
matplotlib
